Create SECURITY.md
It seems you have provided a URL for the "new file" interface in the `develop` branch of the `LedgerHQ/ledger-live` repository. However, since this URL does not point to a specific file, I cannot retrieve or summarize its contents.

If you provide a URL pointing to a specific file, I can retrieve and summarize its content for you.